provide app setting for visual theme selection
