+++ .circleci/config.yml
@@ -64,6 +64,9 @@ jobs:
       - run:
           name: Integration test reviews feature
           command: ./gradlew :features:reviews:featureTesting:testDebugUnitTest --no-daemon --max-workers=2 --stacktrace
+      - run:
+          name: Integration test settings feature
+          command: ./gradlew :features:settings:featureTesting:testDebugUnitTest --no-daemon --max-workers=2 --stacktrace
       - store_artifacts:
           path: app/build/reports
           destination: reports
@@ -71,6 +74,8 @@ jobs:
           path: features/reviews/featureTesting/build/test-results
       - store_test_results:
           path: features/news/featureTesting/build/test-results
+      - store_test_results:
+          path: features/settings/featureTesting/build/test-results
 workflows:
   version: 2
   build_and_test:
