+++ app/build.gradle
@@ -64,6 +64,7 @@ dependencies {
     implementation project(':features:news:di')
     implementation project(':features:reviews:di')
     implementation project(':features:settings:presentation')
+    implementation project(':features:common:presentation')
 
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
