+++ app/src/androidTest/assets/feature/article_sharing_journey.feature
