+++ features/common/navigation/build.gradle
@@ -34,16 +34,15 @@ android {
 }
 
 dependencies {
-    def nav_component_version = "2.3.0"
+    def nav_component_version = "2.3.1"
 
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    implementation 'androidx.core:core-ktx:1.3.1'
+    implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
 
     // Navigation component
     api "androidx.navigation:navigation-fragment-ktx:$nav_component_version"
     api "androidx.navigation:navigation-ui-ktx:$nav_component_version"
-    api "androidx.navigation:navigation-runtime-ktx:$nav_component_version"
 
 }
