+++ features/common/presentation/src/main/res/values-night/styles.xml
@@ -1,0 +1,3 @@
+<resources>
+    <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.MaterialComponents.Dark" />
+</resources>
