+++ features/common/presentation/src/main/res/values/colors.xml
@@ -3,5 +3,4 @@
     <color name="colorPrimary">#6200EE</color>
     <color name="colorPrimaryDark">#3700B3</color>
     <color name="colorAccent">#03DAC5</color>
-    <color name="cardBackground">#FFFFFF</color>
 </resources>
