+++ features/home/presentation/src/main/res/layout/activity_main.xml
@@ -41,7 +41,6 @@
         android:id="@+id/bottom_nav"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@android:color/white"
         app:menu="@menu/bottom_nav" />
 
 </LinearLayout>
