+++ features/news/presentation/src/main/res/layout/news_headline.xml
@@ -28,7 +28,6 @@
         android:layout_marginTop="4dp"
         android:layout_marginBottom="4dp"
         app:cardCornerRadius="2.5dp"
-        android:background="@color/cardBackground"
         android:foreground="?android:attr/selectableItemBackground"
         android:onClick="@{() -> headlineClickListener.invoke(newsTitle)}">
