+++ features/reviews/presentation/src/main/res/layout/movie.xml
@@ -20,7 +20,6 @@
         android:layout_height="wrap_content"
         android:layout_margin="4dp"
         app:cardCornerRadius="2dp"
-        android:background="@color/cardBackground"
         android:foreground="?android:attr/selectableItemBackground"
         android:onClick="@{() -> movieClickListener.invoke(movie)}">
