+++ features/settings/featureTesting/build.gradle
@@ -22,28 +22,38 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
-
-    // Java 8
     compileOptions {
-        sourceCompatibility = 1.8
-        targetCompatibility = 1.8
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = '1.8'
     }
 
-    kotlinOptions {
-        jvmTarget = "1.8"
+    // for robolectric testing
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+            returnDefaultValues = true
+        }
     }
 }
 
 dependencies {
+    def fragment_version = "1.3.0-beta01"
 
-    implementation fileTree(dir: "libs", include: ["*.jar"])
+    // Project modules
+    implementation project(':features:settings:presentation')
+
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'androidx.preference:preference-ktx:1.1.1'
     implementation 'com.google.android.material:material:1.2.1'
-    implementation "androidx.fragment:fragment-ktx:1.3.0-beta01"
+    implementation 'androidx.preference:preference-ktx:1.1.1'
 
-    // Navigation component
-    implementation project(':features:common:navigation')
+    // Fragment testing
+    debugImplementation "androidx.fragment:fragment-testing:$fragment_version"
+
+    testImplementation project(":features:common:featureTesting")
+
 }
