+++ settings.gradle
@@ -1,3 +1,4 @@
+include ':features:settings:featureTesting'
 include ':features:common:domain'
 include ':features:reviews:di'
 include ':features:common:data'
